requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==1.0.0
Django==4.1.3
Pillow==9.3.0
python-decouple==3.6
sqlparse==0.4.3
reportlab==3.4.*
https://github.com/thiagopena/geraldo/archive/master.zip
